I find it necessary to add a timeval to the plugin _close() functions.
In the case of rssm there may be time periods where there are no packets
received.  When this happens the "last-packet-time" was set to far
in the past (or zero).

I think it will be better for the parent to pass the time it calls
the close function.  Also this should make the open-close interval
better for calculating pkts/sec since the close of one interval
will be equal to the open of the next interval and close-sec should
be equal to the -t arg, rather than t-1.