Some changes for portability to Solaris.
On Solairs (10) libresolv includes ns_initparse() and ns_sprintrr()
which were the functions we were checking for in configure.  Solaris
libresolv does not include a useable ns_name_uncompress and p_rcode()
is questionable.  There was an assumption that if the former were
available, the latter were as well.  Now configure checks for all
of those functions and compiles the full dump_dns.c only if they
are all found.

Another little change to the Makefile because Solaris make doesn't
support the -C option.