The RSSM plugin was missing the "rcode-volume" metric defined in the RSSAC 002 document, which was published after the initial RSSM plugin was written.
One complication with RCODEs is that an EDNS0 OPT record can convey an
Extended RCODE.

For this reason, the RSSM plugin now depends on ldns.  If a response
packet has a non-zero ARCOUNT, then it is parsed using ldns_wire2pkt()
and the extended RCODE is extracted using ldns_pkt_edns_extended_rcode().