Made sure dnscap outputs a pcap header when writing to stdout.
When writing to stdout, and when filtering (e.g., by address) it is
possible that no packets will be output.  However, an empty pcap
file is not equivalent to an empty (zero size) file.  When used in
a sequence of pipe commands, downstream pcap readers always expect
to read the pcap header, even if there will be no packets to read.
With this commit, dnscap now calls dumper_open() immediately when
writing to stdout so that the header will be present.