Change how ICMP packets are passed to plugins.
Until now, ICMP and UDP/TCP packets were passed differently.  For UDP/TCP we
pass a pointer to the payload (after the UDP/TCP headers).  But for ICMP we
were passing a pointer to the IP header.  That meant the plugin would have to
again skip over the IP header to get to the ICMP data.

Now for ICMP packets we pass a pointer to the ICMP structure.