A change to the interface of the family of output() functions.
This function has two args that point to packet data.  'pkt_copy'
points to the start of the full, raw packet that includes Ethernet,
IP, and transport headers.  The 'dnspkt' arg pointed to the start
of a DNS message, or may be NULL for non-DNS packets.

A problem arises with plugins that may want to access non-DNS
messages, such as ICMP.  With the old interface, a plugin that wants
ICMP payload would have to start with 'pkt_copy' and re-strip away
the Ethernet/IP headers, replicating code in the main dnscap program.

This change adds an 'isdns' flag to the family of output() functions.
It also renames 'dnspkt' to 'payload'.

When isdns is non-zero, payload points to the start of a DNS message.
It would be the caller's mistake to pass isdns=1 and payload=0.

When isdns is zero, and proto==ICMP, payload points to the start
of the ICMP header.