The rzkeychange data collection plugin has a need to know if an IP address is a "responder" or not, because when dnscap is instructed to collect ICMP with -I, it processes all ICMP packets, not just those limited to responders (or initiators).
To accomplish this there is now an is_responder() function in the main
dnscap code.  Plugins can access this as a "callback" function.

Previously there was just a single logerr() callback function that
was passed to a plugin's plugin_start() function.  Now plugin_start()
takes a structure of callbacks as its argument.  That structure has
pointers to logerr() and to is_responder().  NOTE this change has not
been propagated to the other plugins yet.