Added Sample module
Added a module that allows the output to be sampled. With the new -q <int> option, the user is now able to decide whether the output should be sampled as follows: Only every nth query is output and only answers that correspond to one of these sampled queries will be output.