Fix #24: Handle packet drops
Update `pcap-thread` to v1.2.0 to get the new callback queue mode which
puts that mode into using pthread conditions if all pcaps are offline and
keeps us from losing packets.

Use `pcap_thread_dropback()` callback to get the notification when a
packet was dropped because the queue was full, indicating that we can't
process all the packets.  Added this stats to the `-S` output as total
and per interface as `ptdrop`.  Changed the output for each interface
to not cut of information, for example interface name was cut to
4 characters.