Improved plugin prototyping and more flexible plugin callbacks
Previosly there was no prototyping of functions between the main
dnscap process and the plugin loadable object.  Now the plugin_func_t()
prototypes are propogated to all plugins.

In developing the rzkeychange plugin it was necessary to have a
callback from the plugin to the main process to know whether or not an
IP address was in the configured responders list.  Rather than add a
new arg to the plugin start() function, the "logerr" arg was replaced
with a structure pointer that can reference multiple callback functions,
hopefully future-proofing the API somewhat.