Implement RFC8145 key tag signal collection in rzkeychange plugin
key tag signal collection is enabled with the -k <zone> option to
the rzkeychange plugin.  Signals are saved during the collection
interval, and then sent to the specified <zone>, one at a time,
at the end of the interval.  Only root zone signals are collected.
The format is:

   timestamp.ip-address.key-tag-signal.<zone>

ip-address is inet_ntop() format with '.' and ':' changed to '-'.