daemon: Attempt to drop supplemental groups
This adds a check for initgroups and setgroups and uses them to
drop supplemental groups when dropping privileges.

Prefer initgroups to maintain all the groups of the target user
in the instance that restrictive groups are used (rare) and fallback
to setgroups as most likely we are dropping from a root/super user.